Move layout to common/components
