+++ src_ts/components/common/components/etools-status/etools-status-styles.ts
@@ -1,5 +1,5 @@
 import {html} from '@polymer/polymer';
-import {SharedStyles} from '../../styles/shared-styles';
+import {SharedStyles} from '../../../styles/shared-styles';
 
 export const etoolsStatusStyles = html`
   ${SharedStyles}
