+++ src_ts/components/common/components/etools-status/etools-status.ts
@@ -1,5 +1,5 @@
 import {connect} from 'pwa-helpers/connect-mixin';
-import {store, RootState} from '../../../redux/store';
+import {store, RootState} from '../../../../redux/store';
 import {PolymerElement, html} from '@polymer/polymer';
 import {timeOut} from '@polymer/polymer/lib/utils/async.js';
 import {Debouncer} from '@polymer/polymer/lib/utils/debounce.js';
@@ -8,9 +8,9 @@ import '@polymer/iron-icons/av-icons.js';
 import '@unicef-polymer/etools-content-panel/etools-content-panel.js';
 import {etoolsStatusStyles} from './etools-status-styles';
 import './etools-action-button.js';
-import {StatusAction, Status} from '../../../typings/etools-status.types';
+import {StatusAction, Status} from '../../../../typings/etools-status.types';
 import {property} from '@polymer/decorators';
-import CommonMixin from '../../common/mixins/common-mixin';
+import CommonMixin from '../../mixins/common-mixin';
 
 /**
  * Etools item(partner/agreement/intervention/report etc.) status display element
