+++ src_ts/components/pages/agreements/pages/components/agreement-status.ts
@@ -1,9 +1,9 @@
 import {PolymerElement, html} from '@polymer/polymer';
 import '@polymer/iron-icons/av-icons.js';
 import CONSTANTS from '../../../../../config/app-constants';
-import EtoolsStatusCommonMixin from '../../../../layout/etools-status/etools-status-common-mixin';
-import '../../../../layout/etools-status/etools-status';
-import '../../../../layout/etools-status/etools-status-common-mixin.js';
+import EtoolsStatusCommonMixin from '../../../../common/components/etools-status/etools-status-common-mixin';
+import '../../../../common/components/etools-status/etools-status';
+import '../../../../common/components/etools-status/etools-status-common-mixin.js';
 import {fireEvent} from '../../../../utils/fire-custom-event';
 import {property} from '@polymer/decorators';
 import '../../data/agreement-termination';
