+++ src_ts/components/pages/agreements/pages/details/agreement-details.ts
@@ -20,9 +20,9 @@ import {
 } from '../../../../../redux/actions/upload-status';
 import {store, RootState} from '../../../../../redux/store';
 import {connect} from 'pwa-helpers/connect-mixin';
-import '../../../../layout/etools-form-element-wrapper';
-import '../../../../layout/etools-cp-structure.js';
-import '../../../../layout/year-dropdown.js';
+import '../../../../common/components/etools-form-element-wrapper';
+import '../../../../common/components/etools-cp-structure';
+import '../../../../common/components/year-dropdown.js';
 import pmpEndpoints from '../../../../endpoints/endpoints.js';
 import CONSTANTS from '../../../../../config/app-constants';
 import CommonMixin from '../../../../common/mixins/common-mixin';
@@ -44,7 +44,7 @@ import {isJsonStrMatch} from '../../../../utils/utils';
 import {partnersDropdownDataSelector} from '../../../../../redux/reducers/partners';
 import {fireEvent} from '../../../../utils/fire-custom-event';
 import {property} from '@polymer/decorators';
-import {EtoolsCpStructure} from '../../../../layout/etools-cp-structure';
+import {EtoolsCpStructure} from '../../../../common/components/etools-cp-structure';
 import {MinimalStaffMember} from '../../../../../models/partners.models';
 import {EtoolsDropdownEl} from '@unicef-polymer/etools-dropdown/etools-dropdown';
 import {Agreement, LabelAndValue, PartnerStaffMember} from '@unicef-polymer/etools-types';
