+++ src_ts/components/pages/interventions/interventions-module.ts
@@ -10,16 +10,16 @@ import EndpointsMixin from '../../endpoints/endpoints-mixin';
 import ScrollControlMixin from '../../common/mixins/scroll-control-mixin';
 import ModuleMainElCommonFunctionalityMixin from '../mixins/module-common-mixin';
 import ModuleRoutingMixin from '../mixins/module-routing-mixin';
-import '../../layout/page-content-header.js';
-import '../../layout/page-content-header-slotted-styles.js';
-import '../../layout/etools-error-messages-box.js';
-import '../../layout/etools-tabs.js';
+import '../../common/components/page-content-header.js';
+import '../../styles/page-content-header-slotted-styles.js';
+import '../../common/components/etools-error-messages-box.js';
+import '../../common/components/etools-tabs.js';
 import './data/intervention-item-data.js';
 import '../agreements/data/agreement-item-data.js';
 import {pageLayoutStyles} from '../../styles/page-layout-styles';
 import {SharedStyles} from '../../styles/shared-styles';
 import {buttonsStyles} from '../../styles/buttons-styles';
-import {pageContentHeaderSlottedStyles} from '../../layout/page-content-header-slotted-styles';
+import {pageContentHeaderSlottedStyles} from '../../styles/page-content-header-slotted-styles';
 import {isEmptyObject, isJsonStrMatch} from '../../utils/utils';
 import {Debouncer} from '@polymer/polymer/lib/utils/debounce';
 import {timeOut} from '@polymer/polymer/lib/utils/async';
