+++ src_ts/components/pages/interventions/pages/new/intervention-new.ts
@@ -8,7 +8,7 @@ import {csoPartnersSelector} from '../../../../../redux/reducers/partners';
 import CONSTANTS from '../../../../../config/app-constants';
 import {ColumnStyles} from '../../../../styles/column-styles';
 import {template} from './intervention-new.template';
-import '../../../../layout/etools-form-element-wrapper';
+import '../../../../common/components/etools-form-element-wrapper';
 import '@polymer/paper-toggle-button/paper-toggle-button.js';
 import '@unicef-polymer/etools-info-tooltip/etools-info-tooltip';
 import '@unicef-polymer/etools-date-time/datepicker-lite';
