+++ src_ts/components/pages/partners/components/partner-status.ts
@@ -1,7 +1,7 @@
 import {PolymerElement, html} from '@polymer/polymer';
-import '../../../layout/etools-status/etools-status.js';
+import '../../../common/components/etools-status/etools-status.js';
 import {createDynamicDialog} from '@unicef-polymer/etools-dialog/dynamic-dialog';
-import EtoolsStatusCommonMixin from '../../../layout/etools-status/etools-status-common-mixin';
+import EtoolsStatusCommonMixin from '../../../common/components/etools-status/etools-status-common-mixin';
 import CONSTANTS from '../../../../config/app-constants.js';
 import {isEmptyObject} from '../../../utils/utils';
 import {fireEvent} from '../../../utils/fire-custom-event';
