+++ src_ts/components/pages/partners/pages/details/partner-details.ts
@@ -10,10 +10,10 @@ import RiskRatingMixin from '../../../../common/mixins/risk-rating-mixin.js';
 import '@unicef-polymer/etools-content-panel/etools-content-panel.js';
 import '@unicef-polymer/etools-dropdown/etools-dropdown-multi.js';
 import '@unicef-polymer/etools-data-table/etools-data-table.js';
-import '../../../../layout/etools-form-element-wrapper';
+import '../../../../common/components/etools-form-element-wrapper';
 
-import '../../../../layout/etools-error-messages-box.js';
-import '../../../../layout/icons-actions';
+import '../../../../common/components/etools-error-messages-box.js';
+import '../../../../common/components/icons-actions';
 
 import {pageCommonStyles} from '../../../../styles/page-common-styles';
 import {gridLayoutStyles} from '../../../../styles/grid-layout-styles';
