+++ src_ts/components/pages/reports/pages/summary/sent-bk-comments.ts
@@ -1,6 +1,6 @@
 import {PolymerElement, html} from '@polymer/polymer';
 import '@unicef-polymer/etools-dialog/etools-dialog.js';
-import '../../../../layout/etools-form-element-wrapper';
+import '../../../../common/components/etools-form-element-wrapper';
 import {fireEvent} from '../../../../utils/fire-custom-event';
 import {property} from '@polymer/decorators';
